fix: Better separation of responsive header styles for web and native
